feat: add account errors
